requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.0
idna==3.6
multidict==6.0.4
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
twilio==8.10.2
urllib3==2.1.0
yarl==1.9.3
